requirements.txt
alabaster==0.7.12
amqp==2.6.0
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appdirs==1.4.4
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Automat @ file:///tmp/build/80754af9/automat_1600298431173/work
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.7.1
bidict==0.21.2
billiard==3.6.3.0
bitarray==0.9.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
blis==0.4.1
bokeh==1.2.0
boto==2.49.0
Bottleneck==1.2.1
cairocffi==1.0.2
CairoSVG==2.4.0
camelot-py==0.7.3
catalogue==1.0.0
celery==4.4.4
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
clyent==1.2.2
codecov==2.0.15
colorama==0.4.1
comtypes==1.1.7
conda==4.9.2
conda-build==3.18.8
conda-package-handling==1.3.11
conda-verify==3.4.2
configparser==3.5.3
constantly==15.1.0
contextlib2==0.5.5
coverage==4.5.4
cryptography==2.7
cssselect==1.1.0
cssselect2==0.2.1
cycler==0.10.0
cymem==2.0.3
Cython==0.29.12
cytoolz==0.10.0
dask==2.1.0
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
distributed==2.1.0
distro==1.4.0
Django==3.0.7
django-celery-results==1.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.14
en-core-web-sm==2.2.5
entrypoints==0.3
et-xmlfile==1.0.1
eventlet==0.17.4
excalibur-py==0.4.2
fastcache==1.1.0
filelock==3.0.12
Flask==0.12.2
Flask-Cors @ file:///tmp/build/80754af9/flask-cors_1609957731919/work
Flask-SocketIO==4.3.2
future==0.18.2
gevent==1.4.0
ghostscript==0.6
glob2==0.7
greenlet==0.4.15
gunicorn==18.0
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink @ file:///tmp/build/80754af9/hyperlink_1610130746837/work
idna==2.9
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==1.6.0
incremental==17.5.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-launcher @ file:///tmp/build/80754af9/jupyterlab_launcher_1601490670252/work
jupyterlab-server==1.0.0
keyring==18.0.0
kiwisolver==1.1.0
kombu==4.6.10
lazy-object-proxy==1.4.1
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
lxml==4.3.4
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
murmurhash==1.0.2
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
nose==1.3.7
notebook==6.0.0
numba==0.44.1
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.9.1
odo==0.5.1
olefile==0.46
opencv-python==4.1.0.25
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pdfkit==0.6.1
pdfminer.six==20181108
pdfplumber==0.5.12
pdfquery==0.4.3
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.1.0
pip-review==1.0
pipreqs==0.4.10
pkginfo==1.5.0.1
plac==1.1.3
plaid-python==4.0.0
pluggy==0.12.0
ply==3.11
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
psycopg2==2.7.6.1
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
pycurl==7.43.0.3
pyfiglet==0.8.post1
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest @ file:///tmp/build/80754af9/pyhamcrest_1615748656804/work
PyJWT==1.7.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyPDF2==1.26.0
Pyphen==0.9.5
pyquery==1.4.0
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.7.0
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-cov==2.7.1
pytest-doctestplus==0.3.0
pytest-mpl==0.10
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
pytest-runner==5.1
python-dateutil==2.8.0
python-dotenv==0.15.0
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2020.1
PyWavelets==1.0.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.8.0
redis==3.5.3
requests==2.23.0
roman==3.2
rope==0.14.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
snowballstemmer==1.9.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
spacy==2.2.3
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.5
sqlparse==0.3.1
srsly==2.0.0
statsmodels==0.10.0
sympy==1.4
tables==3.5.2
tabula-py==1.4.1
tabulate==0.8.3
tblib==1.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.3.1
tinycss2==1.0.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
twilio==6.51.0
Twisted==18.9.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
unicodecsv==0.14.1
urllib3==1.25.9
vine==1.3.0
Wand==0.5.6
wasabi==0.6.0
wcwidth==0.1.7
WeasyPrint==48
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wkhtmltopdf==0.2
wrapt==1.11.2
xgboost==0.90
xlrd==1.2.0
XlsxWriter==1.1.8
xlwings==0.15.8
xlwt==1.3.0
yarg==0.1.9
zict==1.0.0
zipp==3.1.0
zope.interface==4.6.0
